requirements.txt
tensorflow-cpu==2.10.0
numpy==1.23.5
h5py==3.7.0
flask
gunicorn
Pillow
reportlab
gdown